src/app/shared/swagger/swagger.service.spec.ts
import { ConnectionBackend, Http, HttpModule } from '@angular/http';
import { TestBed } from '@angular/core/testing';
import { MockBackend, MockConnection } from '@angular/http/testing';
import { ConfigService } from '../config/config.service';
import { SwaggerService } from './swagger.service';
import { SwaggerRequestFactory } from './swagger-request.factory';

describe('SwaggerService tests', () => {
  let service: SwaggerService;
  let requestFactory: SwaggerRequestFactory;
  let configService: ConfigService;
  let backend: MockBackend;
  let http: Http;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpModule],
      declarations: [],
      providers: [
        ConfigService,
        SwaggerService,
        SwaggerRequestFactory,
        Http,
        {
          provide: ConnectionBackend,
          useClass: MockBackend,
        },
      ],
    });
  });

  beforeEach(() => {
    requestFactory = TestBed.get(SwaggerRequestFactory);
    service = TestBed.get(SwaggerService);

    configService = TestBed.get(ConfigService);
    backend = TestBed.get(ConnectionBackend);
    http = TestBed.get(Http);
  });

  it('should be instantiable', () => {
    expect(service).toBeDefined();
  });

  describe('getSwaggerFile', () => {
    it('should fail when input param is missing', () => {
      service.getSwaggerFile(null)
        .subscribe(() => {
          fail();
        }, error => expect(error).toBe('input is required'));
    });

    it('should fail when namespace param is missing', () => {
      service.getSwaggerFile({})
        .subscribe(() => {
          fail();
        }, error => expect(error).toBe('namespace param is required'));
    });

    it('should fail when calling service fails', () => {
      backend.connections.subscribe((connection: MockConnection) => {
        connection.mockError(new Error('error with connection'));
      });

      const request = requestFactory.toGetSwaggerFileRequest('console-server');
      service.getSwaggerFile(request)
        .subscribe(() => {
          fail();
        }, error => expect(error.message).toBe('error with connection'));
    });
  });

});
